WIP: Add read only rootfs support
Trying to get a squashfs and initramfs in to the /boot partition of a
wic image so allow for experimenting with mount the squashfs as the
rootfs.

Needed a custom wic kickstart file and a custom wic source plugin to
work around the squashfs files not being the
`build/tmp/deploy/images/<machine>/` directory when the wic image is
created (that directory doesn't get populated until the
`do_image_complete` task is run which is after `do_image_wic` is run.

Can now boot into initramfs init script and from there switch_root
into the full system (tested on a raspberrypi4 machine).

Signed-off-by: Theodore A. Roth <troth@openavr.org>