wic: Don't use sourceparams
If "loader=u-boot" is set in soruceparams, then u-boot is configured to
use extlinux config files which breaks booting before boot.scr is given
a chance to load.

We want to use boot.scr for boot instructions if it is available instead
of extlinux.conf, so remove `--sourceparams="loader=u-boot"` from the wic
wks file.

Signed-off-by: Theodore A. Roth <troth@openavr.org>