init-boot: Disable dropping shell in boot script
Dropping into a shell for debugging during init boot script is
problematic for the raspberry pi. When booting without a serial console,
the monitor just remains blanked and you do not see the shell prompt.
Thus it appears that the device boot is hung, but it is waiting for
input on the serial console that it will never get.

The option to drop into a shell during the init script is still present,
but is only enabled by adding `debug-openavr-init` to the kernel cmdline
args.

Signed-off-by: Theodore A. Roth <troth@openavr.org>