u-boot: Use boot.scr instead of uEnv.txt
The old uEnv.txt u-boot script was no longer working with the beaglebone
black since the update to honister/main.

Convert to using a boot.scr u-boot script for beaglebone black and AI
boards which actually makes things a bit simpler.

There is still some duplicated code that needs to be cleaned up for
better maintainability.

Signed-off-by: Theodore A. Roth <troth@openavr.org>