Add devcfg squashfs handling
Adds a mechanism for adding device specific, readonly configuration data
to the device image using a `devcfg.squashfs` file.

If the initramfs init script finds the `devcfg.squashfs` in the `/config`
partition filesystem, it will create an overlayfs with it and the root
squashfs, thus mapping device specfic data into the readonly rootfs.

If the devcfg file is not present, then the device with boot with the
generic configuration in the rootfs.

Testing usage:

* Create a devcfg.squashfs file with a given serial number.
* Create a wic image file and load it onto an SD card.
* The devcfg.squashfs image can be loaded in two ways:
  + Copy the file to the SD card `/config` filesystem directly on your
    workstation and then boot the device from the SD card.
  + Boot device from the SD card and then scp the file into place and
    then reboot to have the configuration enabled.

The devcfg image can be updated or removed on the running device.

The next development task is to add a tool to the image to run attempt
to fetch the devcfg image from a server (https/tftp?) and install it on
the device if no devcfg image is present on the device.

Signed-off-by: Theodore A. Roth <troth@openavr.org>