Temporary fix MLO booting on bbai
This commit shall be reverted once the pending meta-ti changes are
merged to master with include the referenced commit in
u-boot-ti-staging.

The commit in u-boot-ti-staging which fixes the actual breakage of MLO
is bae522d534768449e811eca14edae29db618aa06. Once the master branch of
meta-ti includes that commit, this commit canbe reverted.

Signed-off-by: Theodore A. Roth <troth@openavr.org>