u-boot: Set UBOOT_CONFIG_FRAGMENTS
This is the new way that upstream sets config for beaglebone-ai64 and
just builds on top of the config for the j721e_evm_r5_defconfig and
j721e_evm_a72_config machine configuration files. Without this, you only
get the default base evm config.

This really should be fixed in meta-ti.

Signed-off-by: Theodore A. Roth <troth@openavr.org>