init-boot.sh: Rework setting of hostname
Setting the hostname by grepping for Serial in /proc/cpuinfo always
resulted in a sha256sum regardless of if Serial is present, thus giving
a false indication that a serial number is available for the device.

Also, for the beagle-bone AI-64, the serial-number file is present under
the devicetree files, but no serial number is present in /proc/cpuinfo.
This rework makes things work similarly for all the devices I've tried
so far.

Signed-off-by: Theodore A. Roth <troth@openavr.org>